engine: extract dispatch::cards (#159)
